// src/shared/dataprovider/api/model/updateBalanceTransferRequestDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BalanceTransferRequestBreakdownDestinationDTO } from './balanceTransferRequestBreakdownDestinationDTO';
import { BalanceTransferRequestBreakdownSourceDTO } from './balanceTransferRequestBreakdownSourceDTO';
import { TopUpStatusEnum } from './topUpStatusEnum';
import { TransactionTypeEnum } from './transactionTypeEnum';

export interface UpdateBalanceTransferRequestDTO { 
    TransactionNumber?: string;
    TransferType?: number;
    Status?: number;
    Amount?: number;
    Remarks?: string;
}